Add a script that generates property classes from yaml and templates
The property classes are very similar. Defining them all manually is a
lot of work and maintainance. However, with some clever python scripting
we can instead generate them using a structure defined in a YAML file.
This greatly reduces the amount of duplicate code we need to write,
while also giving an improved overview of what the actual data structure
looks like. We can even generate autotests for all the property classes.