Add support for icon styling
Three basic attributes:

- Color
- Size
- Source

A theme might want to apply certain colors to controls that use specific icons.
For example, changing the color of an icon for a spinbox's increment button when pressed.
Currently it just reads a string into QColor::fromString to get the color.

A theme might want to set default icon sizes for certain controls.
For example, push buttons with 16x16 icons and toolbar buttons with 22x22 icons.
You can specify size in different ways:

- Use constant integer or floating point values.
- Use string preset values. Presets: small (16), small-medium (22), medium (32), large (48), huge (64), enormous (128)
- Set width and height with one value. Example YAML: `size: small`
- Set width and height separately. Example YAML: `size: {width: small, height: small}`

A theme might want to specify a theme icon name or icon file for certain kinds of controls.
For example, the increment and decrement buttons of a spinbox.
The source is an std::variant that can be a QString for theme icon names or a QUrl for icon files.
There is also a fallback system so that the example spinbox could fall back to different icons
when requested icons are not in the user's icon theme. Example YAML:

- Single value: `source: go-up`
- List with fallbacks: `source: [go-up, alternative-icon-name, /path/to/icon/dir/go-up.png]`