Add icon support to data model
Three basic attributes:

- Color
- Size
- Source

A theme might want to apply certain colors to controls that use specific icons.
For example, changing the color of an icon for a spinbox's increment button when pressed.

A theme might want to set default icon sizes for certain controls.
For example, push buttons with 16x16 icons and toolbar buttons with 22x22 icons.

A theme might want to specify a theme icon name or icon file for certain kinds of controls.
For example, the increment and decrement buttons of a spinbox.
The source is represented by a theme icon name variable and a file url variable. Only set the one you want to use.