output/quick: Force creation of a Style attached object for custom Elements
If we specify properties that result in an attached Element being
treated as an explicit element type, ensure we also create a Style
object for said element. This ensures the Style and any children are
using the Element to query which styles to apply, rather than Style
being allowed to inherit from an element higher up in the hierarchy.