output/quick: Expose Alignment as a namespace to QtQuick
This exposes the values of Union::Properties::Alignment so QtQuick can
make use of them. Unfortunately using foreign types to expose the enum
directly did not work.

output/quick: Add conversion functions to convert Alignment to Qt alignment

So that we don't need to repeat this code in multiple places where we
want to set Qt alignment based on Union alignment values.