tools/properties: Use QJSValue for getters of values
These values are optional on the C++ side. We should communicate the
same to QML so things fall back to QtQuick default behaviors.
Unfortunately there is no "optional<>" support in the QtQuick enigne.
However, we can use QJSValue ourselves and use QJSValue(Undefined) to
indicate a value is not set.