output/quick: Prevent polish loops in PositionerLayout
If we trigger a polish while layouting we can end up in polish loops
when a control is used inside a QtQuick.Layout. To avoid this, don't
call polish while we're layouting. Since we still need to trigger a
polish due to layouted controls changing while we're layouting, queue a
polish once we're done with layouting and we've detected that we need
another round.