output/kirigami: Disable inherit for PlatformTheme by default for now
PlatformTheme's inherit uses a different hierarchy from Union. This
means that sometimes it will be inheriting from an item that uses
different colors than the current item should be using. The solution is
to somehow inject non-inherting PlatformTheme instances for elements
that shouldn't inherit, but that's tricky to solve. So for now, just
default inherit to false so we always read our own colors.