Allow recoloring background-image with help of mask
This makes background-image property to allow `url mask|inverted-mask color` entry.

If user adds url, mask and color, then the svg or other image in the url is used as a mask, and
the color is used for recoloring the resulted texture.

This also adds a pyproject etc info to the property generator with uv