output/quick: Size corners to only one corner size, not both
We don't need to account for other corners' sizes since the geometry of
a single corner should be enough for that corner and its border/outline.
Accounting for other corners actually leads to issues because we end up
accounting for the corner size of corner that's much bigger which leads
to rendering glitches.