output/quick: Position items in PositionerLayout after updating padding
Updating an items' padding may cause Control to shift around things
inside the contentItem, which means our calculated positions need to be
updated based on the new contentItem position. Rather than needing a
full round of layouting for this, we can just place the positioned items
after we've updated padding, so that positioning can account for the
changed positions immediately.

Fixes #63